refactor: 버튼 리펙토링 #23
