+++ src/components/common/Button/index.ts
@@ -1,0 +1,2 @@
+export { default as Button } from './Button';
+export { default as ButtonWrapper } from './ButtonWrapper';
